Core/PacketIO: Update opcodes for 11.0.5
Port From (https://github.com/TrinityCore/TrinityCore/commit/ac7a95b45ca239d91b3daa8e2f77376382b092d2)

+++ Source/Game/Networking/Packet.cs
@@ -96,7 +96,7 @@ namespace Game.Networking
 
         public WorldPacket(byte[] data) : base(data)
         {
-            opcode = ReadUInt16();
+            opcode = ReadUInt32();
         }
 
         public ObjectGuid ReadPackedGuid()
